refactor(version): use dynamic versioning
- delete __about__.py
- delete __title__ and __description__
- move __version__ to boox/__init__.py
- delte redundant tests
- make version dynamic in pyproject.toml
- make __version__ the single source of truth
