+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pybooxdrop"
-version = "0.1.3"
+dynamic = ["version"]
 description = "A friendly Python wrapper for the BOOXDrop API — unofficial, but built with care."
 readme = "README.md"
 license = "MIT"
@@ -47,6 +47,9 @@ dev = [
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.version]
+path = "src/boox/__init__.py"
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/boox"]
