--- src/boox/__about__.py
@@ -1,3 +1,0 @@
-__version__ = "0.1.3"
-__title__ = "pybooxdrop"
-__description__ = "A friendly Python wrapper for the BOOXDrop API — unofficial, but built with care."
