+++ src/boox/__init__.py
@@ -2,3 +2,4 @@ from boox.core import Boox
 from boox.models.enums import BooxUrl
 
 __all__ = ["Boox", "BooxUrl"]
+__version__ = "0.1.3"
