+++ tests/test_init.py
@@ -1,0 +1,5 @@
+from boox import __version__
+
+
+def test_version_is_set():
+    assert __version__, "version is not filled"
